refactor: better code readability
