+++ lib/rag/common/errors.ts
