+++ lib/rag/retrieval/status-helpers.ts
